Pass pixel ID/access token through to server-side events
Server-side events are executed in a separate/async process, which
means that they no longer have the context of the current page request.
This will cause a problem if the pixel ID and/or access token
filter hooks are dependant upon information in the current page request
(e.g. the logged in user, post meta, etc.). To remedy this, we
get the pixel ID and access token values at the time of the current
page request and save them to the server-side event data. When the
event is sent up on the server-side, it uses the saved pixel ID
and access token.

It is written to be backward compatible. Anybody that is already
using the library and is manually calling any of the methods
(`FacebookServerSideEvent::send()`, for example) will not be
affected and everything will continue to work.